+++ src/Audit.php
@@ -85,7 +85,7 @@ class Audit extends Plugin
     /**
      * @var string
      */
-    public $schemaVersion = '1.0.2';
+    public $schemaVersion = '1.0.3';
 
     /**
      * @inheritdoc